requirement.txt




sqlalchemy
sqlalchemy.orm
datetime
networkx
matplotlib
fastapi.middleware
uvicorn

pydantic

face_recognition
requests
numpy